+++ e2e/setup-files-after-env-config/setupAsyncFunction.js
@@ -1,0 +1,10 @@
+globalThis.afterEnvAsyncFunctionFinished = false;
+
+module.exports = async () => {
+  await new Promise(resolve =>
+    setTimeout(() => {
+      globalThis.afterEnvAsyncFunctionFinished = true;
+      resolve();
+    }, 2000),
+  );
+};