+++ packages/jest-jasmine2/src/index.ts
@@ -180,7 +180,10 @@ export default async function jasmine2(
     if (esm) {
       await runtime.unstable_importModule(path);
     } else {
-      runtime.requireModule(path);
+      const setupFile = runtime.requireModule(path);
+      if (typeof setupFile === 'function') {
+        await setupFile();
+      }
     }
   }